+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+- ApplicationController to a sigleton (ApplicationScoped), with eager behavior.
+
 ## 1.0.0
 
 ### Added
@@ -26,3 +31,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Result view: Where the QRCode is shown.
 - Unlock view: Where the user will type the SecureCode password to unlock the redirect.
 - 404 and 500 error handlers.
+
+[Unreleased]: https://github.com/JayBIOS/quickresponse/compare/v1.0.0...HEAD